Conflicts:
	SoA/App.cpp
	SoA/Camera.cpp
	SoA/Chunk.h
	SoA/ChunkGenerator.cpp
	SoA/GamePlayScreenEvents.hpp
	SoA/GenerateTask.h
	SoA/LoadTaskShaders.h
	SoA/SOA.vcxproj.filters
	SoA/SpaceSystem.h
	Vorb

+++ SoA/BlockTextureMethods.cpp
@@ -1,7 +1,8 @@
 #include "stdafx.h"
+#include "BlockTextureMethods.h"
 
 #include "BlockData.h"
-#include "BlockTextureMethods.h"
+#include "Chunk.h"
 #include "ChunkMesh.h"
 #include "RenderTask.h"
 #include "VoxelBits.h"
@@ -12,10 +13,10 @@ void BlockTextureMethods::getRandomTextureIndex(BlockTextureMethodParams& params
     //TODO: MurmurHash3
     const MesherInfo* mi = params.mesherInfo;
     const BlockTextureLayer* blockTexInfo = params.blockTexInfo;
-    int seed = getPositionSeed(mi->x + mi->position.x, mi->y + mi->position.y, mi->z + mi->position.z);
+    i32 seed = getPositionSeed(mi->x + mi->position.x, mi->y + mi->position.y, mi->z + mi->position.z);
 
-    float r = (PseudoRand(seed) + 1.0) * 0.5 * blockTexInfo->totalWeight;
-    float totalWeight = 0;
+    f32 r = (f32)((PseudoRand(seed) + 1.0) * 0.5 * blockTexInfo->totalWeight);
+    f32 totalWeight = 0;
 
     // TODO(Ben): Binary search?
     if (blockTexInfo->weights.getLength()) {
@@ -40,10 +41,10 @@ void BlockTextureMethods::getRandomTextureIndex(BlockTextureMethodParams& params
 void BlockTextureMethods::getFloraTextureIndex(BlockTextureMethodParams& params, int& result) {
     //TODO: MurmurHash3
     const MesherInfo* mi = params.mesherInfo;
-    int seed = getPositionSeed(mi->x + mi->position.x, mi->y + mi->position.y, mi->z + mi->position.z);
+    i32 seed = getPositionSeed(mi->x + mi->position.x, mi->y + mi->position.y, mi->z + mi->position.z);
 
-    float r = (PseudoRand(seed) + 1.0) * 0.5 * params.blockTexInfo->totalWeight;
-    float totalWeight = 0;
+    f32 r = (f32)((PseudoRand(seed) + 1.0) * 0.5 * params.blockTexInfo->totalWeight);
+    f32 totalWeight = 0;
 
     const BlockTextureLayer* blockTexInfo = params.blockTexInfo;
     const ui16* tertiaryData = mi->tertiaryData;
